pyproject.toml
[tool.poetry]
name = "chainlit-gemini"
version = "0.1.0"
description = "Chainlit UI for chatting with Gemini 1.5 Pro using Open Router"
authors = ["a-zuev-pc <azuev@outlook.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
chainlit = "^1.1.402"
openai = "^1.41.0"
langchain = "^0.2.14"
pyyaml = "^6.0.2"
langchain-community = "^0.2.12"
langchain-openai = "^0.1.22"
google-generativeai = "^0.7.2"
langchain-google-genai = "^1.0.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
